feat(ui): add responsive desgin for email-verify page
