+++ src/app/[locale]/auth/verify-email/layout.tsx
