+++ src/app/[locale]/auth/verify-email/loading.tsx
